refactor: padronização dos arquivos
