+++ src/app/home/page.tsx
@@ -1,19 +1,19 @@
-"use client"
-import FaqQuestion from "@/components/FaqQuestion"
-import { TestimonialsSection } from "@/components/testimonials-section"
-import { HomeBanner } from "@/components/UI/home/home-banner"
-import { ImpactSection } from "@/components/UI/home/impact-section"
-import { SponsorsSection } from "@/components/UI/home/sponsors-section"
+"use client";
+
+import { Faq } from "@/components/faq";
+import { TestimonialsSection } from "@/components/testimonials-section";
+import { HomeBanner } from "@/components/UI/home/home-banner";
+import { ImpactSection } from "@/components/UI/home/impact-section";
+import { SponsorsSection } from "@/components/UI/home/sponsors-section";
 
 export default function Home() {
-
-  return(
+  return (
     <>
       <HomeBanner />
       <ImpactSection />
       <SponsorsSection />
-      <FaqQuestion />
+      <Faq />
       <TestimonialsSection />
     </>
-  )
+  );
 }
