+++ src/components/accordion.tsx
@@ -1,15 +1,15 @@
-import { Accordion, AccordionDetails, AccordionSummary, Typography, useTheme } from "@mui/material";
+import { AccordionDetails, AccordionSummary, Accordion as MuiAccordion, Typography, useTheme } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 
-interface AccordionFaqProps { 
+interface AccordionProps { 
   question: string;
   answer: string;
 }
 
-export default function AccordionFaq({ question, answer }: AccordionFaqProps) {
+export function Accordion({ question, answer }: AccordionProps) {
   const theme = useTheme();
   return (
-    <Accordion
+    <MuiAccordion
       sx={{
         border: "1px solid",
         mb: "8px",
@@ -19,7 +19,9 @@ export default function AccordionFaq({ question, answer }: AccordionFaqProps) {
       }}
     >
       <AccordionSummary
-        expandIcon={<ExpandMoreIcon sx={{ color: theme.palette.bgColor?.main }} />}
+        expandIcon={
+          <ExpandMoreIcon sx={{ color: theme.palette.bgColor?.main }} />
+        }
         aria-controls="panel1-content"
         id="panel1-header"
         sx={{
@@ -28,11 +30,9 @@ export default function AccordionFaq({ question, answer }: AccordionFaqProps) {
           boxShadow: "0em 0.2em 0.4em rgb(44 44 44 / 40%)",
         }}
       >
-        <Typography component="span">{ question }</Typography>
+        <Typography component="span">{question}</Typography>
       </AccordionSummary>
-      <AccordionDetails>
-        { answer }
-      </AccordionDetails>
-    </Accordion>
+      <AccordionDetails>{answer}</AccordionDetails>
+    </MuiAccordion>
   );
 }
